pyproject.toml
[tool.poetry]
name = "repl_nix_Job-Portal-Web-Application"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
consultor = "^0.2.0"
Flask-SQLAlchemy = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
